Move report to separate page
